Fix parsing of unquoted urns
The structued logs emitted by terraform have `key=value` pairs in the log message. Some characters, for example `$` cause the value to be quoted.

The parsing for the quoted urns worked nicely, however since the log message has a trailing `\n`, if the urn is not quoted, and it is the last key=value of the message, we would include the trailing `\n` as part of the urn.

The fix is to strip the trailing `\n` before attempting any parsing.

This lead to some display issues in https://github.com/pulumi/pulumi/issues/20269. I will leave that issue open, because the CLI should gracefully display urns with control characters in them (there are not realy restrictions as to what shows up in an urn).