Fix secret value logged when type checking fails (#3159)
In the typechecker we have to unwrap the secret value to do the nested
type checking, but if it fails we log the value for the user. While it
is useful to see the actual data to know what shape the data is in, we
have to prevent secrets from being logged. They will still see the
`type`, they just won't be able to see the actual value.

fixes #3158