Remove info.SkipExamples (#3153)
This pull request removes the experimental feature of info.SkipExamples.

It was intended as a more granular example-skipping feature for use on
individual resources in providers.

Only two providers are using this currently, and using the feature
results in loss of documentation. For more context, see
https://github.com/pulumi/pulumi-aws/issues/5691.

Blocked by https://github.com/pulumi/pulumi-aws/issues/5691,
https://github.com/pulumi/pulumi-alicloud/issues/1044.

Fixes #3152.