Remove line wrapper for rawDoc and adjust tests (#3164)
This pull request removes the `genRawDocComment` line wrapper function,
which was used to add newlines to inline doc comments, specifically
`prop.rawdoc` strings, which originate directly from the TF schema (not
the upstream markdown docs). These docs, like their markdown siblings,
are read into the schema and then used to document the SDK.

These added newlines seem to serve no other purpose than to render the
generated SDK files with newlines in the docs.

This change should be made because:
- Most of these doc strings aren't that long as to render the comment
useless
- it is otherwise unnecessary: the Pulumi registry handles its own line
wrapping, and so doo do IDEs that use Intellisense to pull down docs
descriptions for fields and functions
- The schema is simplified, losing a bunch of extra newline characters
- Any docs formatting should be done during SDK gen, not schema gen

This change is a prerequisite/nice-to-have for
https://github.com/pulumi/pulumi-terraform-bridge/issues/1918, because
it allows us to mark up doc strings with necessary HTML for registry
rendering and variable inflection.