Fix panic when gathering attribute (#3178)
We set the TF Dynamic type to nil so that we can handle it later:
https://github.com/pulumi/pulumi-terraform-bridge/blob/main/pkg/tfgen/generate.go#L547.

This means we have to ensure that we perform a nil check on the
attribute being processed; we did so in one location but missed this
one. This PR adds the missing check, plus a test to ensure the provider
will not panic.

Add missing nil check when gathering attributes
Add test for panic

Fixes https://github.com/pulumi/pulumi-terraform-provider/issues/88.