Automatically infer an id for pf resources (#3193)
This applies the logic from `dynamic` to `pf` based resources. As more
and more providers are moved to creating new resources with the plugin
framework, more and more of our upgrades are failing due to the missing
id error. This requires manual work for the team to create a `ComputeID`
mapping.

Since this `id` is not actually used for anything in Pulumi, we have
decided to fallback to hardcoding a "missing ID" value. This will reduce
the number of provider upgrades that require manual intervention by the
team and decrease the delay in releasing upgrades.

closes #3182