Use g.pkg.Name() for determining location of schema file (#3199)
Basically what it says on the box.

When we read the schema into the language filter, it expects a location
in the Pulumi repo. That repo is named after the Pulumi package name,
not the ProviderInfo.Name (which is the Terraform name).
This pull request uses g.pkg.Name() to pass along this information.

Fixes #3198.