Drop schema-tools dependency (#3287)
The bridge is stuck on a very old version of schema-tools due to
https://github.com/pulumi/schema-tools/pull/36 which made the project
only consumable as a CLI. This causes problems when trying to bump
dependencies (e.g. https://github.com/pulumi/pulumi-azuread/pull/2155).

We only use a couple types related to stat tracking. Rather than expose
this as a public package on stats-tools, it's better to simply vendor
what we need.